Refine: Gallery improve & TweetInput refactor
