+++ src/components/Tweet/TweetImage.svelte
@@ -2,4 +2,4 @@
     export let src = ''
 </script>
 
-<img src={src} alt='' class="rounded-2xl">
+<img src={src} alt='' class="rounded-2xl mt-2">
